SQLProxy - Development 3
Encryption and decryption successfully implemented on sqlQuery class.  Decryption is only carried out if the query has not extended to include an SQL keyword after encryption.  This is checked by adding the "add" functionality to the class, which will be removed once this mitigation process has been successfully integrated with the web applications server.